+++ src/library/Attempt/Outcome/Failure/index.ts
@@ -2,7 +2,7 @@
 
 import type {
   ActionableError,
-} from '../../error';
+} from '../../error/exports';
 
 import type {
   DiscriminableOutcome,
