+++ src/library/Attempt/Outcome/Transformer.ts
@@ -1,6 +1,6 @@
 import type {
   ActionableError,
-} from '../error';
+} from '../error/exports';
 
 import type {
   Attempt_Failure_Transformer,
