+++ src/library/Attempt/factory/asynchronous.ts
@@ -10,7 +10,7 @@ import {
 
 import type {
   ActionableError,
-} from '../error';
+} from '../error/exports';
 
 import {
   sanctionedAsync,
